refactor: extract filters into dedicated files
