+++ packages/web/src/filtering/filterByClades.ts
@@ -1,0 +1,16 @@
+import { splitFilterString } from 'src/filtering/splitFilterString'
+import { SequenceAnylysisState } from 'src/state/algorithm/algorithm.state'
+import { formatClades } from 'src/helpers/formatClades'
+
+export function filterByClades(cladesFilter: string) {
+  const cladesFilters = splitFilterString(cladesFilter)
+
+  return (result: SequenceAnylysisState) => {
+    if (!result?.result) {
+      return false
+    }
+
+    const { cladeStr } = formatClades(result.result.clades)
+    return cladesFilters.some((filter) => cladeStr.startsWith(filter))
+  }
+}
