+++ packages/web/src/filtering/filterByQCIssues.ts
@@ -1,0 +1,10 @@
+import { SequenceAnylysisState } from 'src/state/algorithm/algorithm.state'
+
+export function filterByQCIssues(hasNoQcIssuesFilter: boolean, hasQcIssuesFilter: boolean, hasErrorsFilter: boolean) {
+  return ({ result, errors }: SequenceAnylysisState) => {
+    const hasErrors = errors.length > 0
+    const hasIssues = hasErrors || (result && result.diagnostics.flags.length > 0)
+    const hasNoIssues = result && result.diagnostics.flags.length === 0
+    return (hasNoQcIssuesFilter && hasNoIssues) || (hasQcIssuesFilter && hasIssues) || (hasErrorsFilter && hasErrors)
+  }
+}
