+++ packages/web/src/filtering/splitFilterString.ts
@@ -1,0 +1,8 @@
+const FILTER_DELIMITERS = /[\n\r,;]/gi
+
+export function splitFilterString(filter: string) {
+  return filter
+    .split(FILTER_DELIMITERS)
+    .map((f) => f.trim())
+    .filter((f) => f !== '')
+}
